# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenAccounting API
  version: "1.0.0"
  description: |
    REST API for the OpenAccounting accounting platform.

    ## Authentication
    All `/api/v1/...` endpoints require a Bearer token:
    ```
    Authorization: Bearer oa_live_…
    ```

    ## Idempotency
    `POST` endpoints accept an `Idempotency-Key` header. Replays
    within 24 hours return the original response without re-executing
    the mutation. Keys are scoped per token + ledger.

    ## Pagination
    List endpoints return `{"data": [...]}` with a `Link: <…>; rel="next"`
    header when more results are available. Pass `?limit=` (default 50,
    max 200) and `?cursor=` from the previous `Link` header.

    ## Errors
    All errors return RFC 7807 problem-details JSON
    (`Content-Type: application/problem+json`).

    ## Incoming Events
    External systems can push signed events to
    `POST /api/events/{ledger_id}` using the ledger's
    `incoming_events_secret` (HMAC-SHA256 in the
    `X-OA-Event-Signature: sha256=<hex>` header).

    ## Automation Rules
    Ledger owners create rules via the web UI at
    `/ledgers/{id}/automations`. Rules map an allowlisted trigger
    (event type) + conditions → an action (webhook / email /
    categorise transaction) and execute on the jobs queue.
servers:
  - url: /
    description: Same origin
paths:
...
  /api/events/{ledger_id}:
    post:
      operationId: intakeEvent
      summary: Signed incoming event intake
      tags: [Events]
      security: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [external.sync, external.document.received, external.payment.received]
                data:
                  type: object
      responses:
        "202":
          description: Event accepted and queued for rule evaluation
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:
                    type: integer
                  event_type:
                    type: string
        "401":
          description: Missing or invalid X-OA-Event-Signature
...
components:
  parameters:
    LedgerIdPath:
      name: ledger_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...